refactoring(ViewModel, Retrofit..) - ViewModel 도입 - Retrofit 및 data class 수정 (특히 distance부분 컨트롤)
추가적인 리팩토링 필요성
- Koin 도입
- adapter 연결 및 데이터 선택에 따른 뷰의 변화

+++ app/src/main/java/com/dongldh/fike/adapter/MainBindingAdapter.kt
@@ -8,9 +8,9 @@ import android.widget.TextView
 import androidx.databinding.BindingAdapter
 
 @BindingAdapter(value = ["bind:distance", "bind:remain"])
-fun bindDistanceAndRemains(view: TextView, distance: Long, remain: Int) {
+fun bindDistanceAndRemains(view: TextView, distance: Int, remain: Int) {
     val string = "xm / y대"
-    val newString = string.replace("x", distance.toInt().toString()).replace("y", remain.toString())
+    val newString = string.replace("x", distance.toString()).replace("y", remain.toString())
 
     val spannable = SpannableString(newString)
     val slash = newString.indexOf("/")
